+++ autorag/embedding/base.py
@@ -9,26 +9,9 @@ from llama_index.embeddings.openai import OpenAIEmbeddingModelType
 from langchain_openai.embeddings import OpenAIEmbeddings
 from llama_index.embeddings.huggingface import HuggingFaceEmbedding
 
+from autorag import LazyInit
+
 logger = logging.getLogger("AutoRAG")
-
-
-class LazyInit:
-	def __init__(self, factory, *args, **kwargs):
-		self._factory = factory
-		self._args = args
-		self._kwargs = kwargs
-		self._instance = None
-
-	def __call__(self):
-		if self._instance is None:
-			self._instance = self._factory(*self._args, **self._kwargs)
-		return self._instance
-
-	def __getattr__(self, name):
-		if self._instance is None:
-			self._instance = self._factory(*self._args, **self._kwargs)
-		return getattr(self._instance, name)
-
 
 class MockEmbeddingRandom(MockEmbedding):
 	"""Mock embedding with random vectors."""
@@ -37,43 +20,45 @@ class MockEmbeddingRandom(MockEmbedding):
 		return [random() for _ in range(self.embed_dim)]
 
 
+embedding_models = {
+    # llama index
+    "openai": LazyInit(
+        OpenAIEmbedding
+    ),  # default model is OpenAIEmbeddingModelType.TEXT_EMBED_ADA_002
+    "openai_embed_3_large": LazyInit(
+        OpenAIEmbedding, model_name=OpenAIEmbeddingModelType.TEXT_EMBED_3_LARGE
+    ),
+    "openai_embed_3_small": LazyInit(
+        OpenAIEmbedding, model_name=OpenAIEmbeddingModelType.TEXT_EMBED_3_SMALL
+    ),
+    "mock": LazyInit(MockEmbeddingRandom, embed_dim=768),
+    # langchain
+    "openai_langchain": LazyInit(OpenAIEmbeddings),
+    # local model
+    "huggingface_baai_bge_small": LazyInit(
+        HuggingFaceEmbedding, model_name="BAAI/bge-small-en-v1.5"
+    ),
+    "huggingface_cointegrated_rubert_tiny2": LazyInit(
+        HuggingFaceEmbedding, model_name="cointegrated/rubert-tiny2"
+    ),
+    "huggingface_all_mpnet_base_v2": LazyInit(
+        HuggingFaceEmbedding,
+        model_name="sentence-transformers/all-mpnet-base-v2",
+        max_length=512,
+    ),
+    "huggingface_bge_m3": LazyInit(HuggingFaceEmbedding, model_name="BAAI/bge-m3"),
+    "huggingface_multilingual_e5_large": LazyInit(
+        HuggingFaceEmbedding, model_name="intfloat/multilingual-e5-large-instruct"
+    ),
+}
+
+
 class EmbeddingModel:
-	embedding_models = {
-		# llama index
-		"openai": LazyInit(
-			OpenAIEmbedding
-		),  # default model is OpenAIEmbeddingModelType.TEXT_EMBED_ADA_002
-		"openai_embed_3_large": LazyInit(
-			OpenAIEmbedding, model_name=OpenAIEmbeddingModelType.TEXT_EMBED_3_LARGE
-		),
-		"openai_embed_3_small": LazyInit(
-			OpenAIEmbedding, model_name=OpenAIEmbeddingModelType.TEXT_EMBED_3_SMALL
-		),
-		"mock": LazyInit(MockEmbeddingRandom, embed_dim=768),
-		# langchain
-		"openai_langchain": LazyInit(OpenAIEmbeddings),
-		# local model
-		"huggingface_baai_bge_small": LazyInit(
-			HuggingFaceEmbedding, model_name="BAAI/bge-small-en-v1.5"
-		),
-		"huggingface_cointegrated_rubert_tiny2": LazyInit(
-			HuggingFaceEmbedding, model_name="cointegrated/rubert-tiny2"
-		),
-		"huggingface_all_mpnet_base_v2": LazyInit(
-			HuggingFaceEmbedding,
-			model_name="sentence-transformers/all-mpnet-base-v2",
-			max_length=512,
-		),
-		"huggingface_bge_m3": LazyInit(HuggingFaceEmbedding, model_name="BAAI/bge-m3"),
-		"huggingface_multilingual_e5_large": LazyInit(
-			HuggingFaceEmbedding, model_name="intfloat/multilingual-e5-large-instruct"
-		),
-	}
 
 	@staticmethod
 	def load(name: str = ""):
 		try:
-			return EmbeddingModel.embedding_models[name]()
+			return embedding_models[name]()
 		except KeyError:
 			raise ValueError(f"Embedding model '{name}' is not supported")
